# lib/api-zod/src/generated/types/adminMissionData.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * CyberQuest AI defensive training API
 * OpenAPI spec version: 0.1.0
 */
import type { MissionAdmin } from './missionAdmin';

export interface AdminMissionData {
  mission: MissionAdmin;
}