# lib/api-zod/src/generated/types/adminMissionResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * CyberQuest AI defensive training API
 * OpenAPI spec version: 0.1.0
 */
import type { AdminMissionData } from './adminMissionData';

export interface AdminMissionResponse {
  success: true;
  data: AdminMissionData;
}