# lib/api-zod/src/generated/types/attemptInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * CyberQuest AI defensive training API
 * OpenAPI spec version: 0.1.0
 */
import type { Assessment } from './assessment';

export interface AttemptInput {
  assessment: Assessment;
  /**
     * @maxItems 8
     * @items.minLength 2
     * @items.maxLength 60
     */
  findings: string[];
}