# lib/api-zod/src/generated/types/emailScenario.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * CyberQuest AI defensive training API
 * OpenAPI spec version: 0.1.0
 */

export interface EmailScenario {
  senderName: string;
  senderEmail: string;
  recipientName: string;
  subject: string;
  receivedAt: string;
  body: string;
  /** @nullable */
  displayedUrl: string | null;
  /** @nullable */
  attachment: string | null;
}