# lib/api-zod/src/generated/types/successResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * CyberQuest AI defensive training API
 * OpenAPI spec version: 0.1.0
 */
import type { SuccessResponseData } from './successResponseData';

export interface SuccessResponse {
  success: true;
  data: SuccessResponseData;
}